Remove useless imports. Remove some warnings. Rename PA_SCAComponentRepresentative* -> SCAPAComponentRepresentative*. Format.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/SHORT_ComponentSCAControllers@18751 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Tests/functionalTests/component/sca/control/TestSCAPropertyController.java
@@ -187,8 +187,8 @@ public class TestSCAPropertyController extends SCAComponentTest {
 
         scac.setValue("x4", (short) 1);
         assertEquals((short) 1, scac.getValue("x4"));
-        scac.setValue("x5", (int) 1);
-        assertEquals((int) 1, scac.getValue("x5"));
+        scac.setValue("x5", 1);
+        assertEquals(1, scac.getValue("x5"));
         scac.setValue("x6", (long) 1);
         assertEquals((long) 1, scac.getValue("x6"));
